+++ R/hooks.R
@@ -10,6 +10,10 @@
   source = .out.hook, output = .out.hook, warning = .out.hook,
   message = .out.hook, error = .out.hook, plot = .plot.hook,
   inline = .inline.hook, chunk = .out.hook, text = identity,
+  evaluate.inline = function(code, envir = knit_global()) {
+    v = withVisible(eval(parse_only(code), envir = envir))
+    if (v$visible) knit_print(v$value, inline = TRUE, options = opts_chunk$get())
+  },
   evaluate = function(...) evaluate::evaluate(...), document = identity
 )